move to resourcecontainer.artwork package
