+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/data/workbook/Workbook.kt
@@ -18,7 +18,7 @@
  */
 package org.wycliffeassociates.otter.common.data.workbook
 
-import org.wycliffeassociates.otter.common.domain.resourcecontainer.ArtworkAccessor
+import org.wycliffeassociates.otter.common.domain.resourcecontainer.artwork.ArtworkAccessor
 import org.wycliffeassociates.otter.common.domain.resourcecontainer.SourceAudioAccessor
 import java.util.*
 import org.wycliffeassociates.otter.common.persistence.IDirectoryProvider
